优化

+++ Example/ChineseGirl/Class/Model/CGShuoShuo.h
@@ -28,6 +28,6 @@
 @property(nonatomic,assign)BOOL isLike;   //点赞
 +(CGShuoShuo *)getTableWithId:(NSString *)ids;
 + (instancetype)modelWithDic:(NSDictionary *)dic;
-+(NSMutableArray *)reloadTableWithId:(NSInteger)ids;
++(NSMutableArray *)reloadTableWithId:(int)ids;
 +(void)reloadTableRondom;//随机排序说说内容，主要用在首页数据
 @end
